pyproject.toml
[tool.poetry]
name = "radioapi"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.6"
djangorestframework = "^3.14.0"
beautifulsoup4 = "^4.11.2"
requests = "^2.28.2"
lxml = "^4.9.2"
alive-progress = "^3.0.1"
django-rest-swagger = "^2.2.0"
python-dotenv = "^1.0.0"
psycopg2 = "^2.9.9"
celery = "^5.3.6"
flower = "^2.0.1"
redis = "^5.0.1"
drf-yasg = "^1.21.7"
coreapi = "^2.3.3"


[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
isort = "^5.13.2"

[tool.black]
line-length = 120
target-version = ['py37']
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.env
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | node_modules
)/
'''

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
